go back to int milliseconds timer code. hand-picked from the following commits in int_timer_goback/coop_part2_rogue_to_xatrix_merge branches :
	0b624fe41ab6e2170196721b8a3c99df659b538a
	0f731c91521e00b552d818cb27727f9dc4fb6303
	c96213edf77121182d80bc8062ed8e819b576f58
	82cbf4671f5e8ba45bbd2c9ce737abf7dd45a1fc
	63244814685896180a47e05077daa29d6beec421
	1cbba64f72cc7ae3e23c4cb594afe4c45436b7e4

+++ client/cl_scrn.c
@@ -805,8 +805,8 @@ int entitycmpfnc( const entity_t *a, const entity_t *b )
 void SCR_TimeRefresh_f (void)
 {
 	int		i;
-	double		start, stop;
-	double	time;
+	int		start, stop;
+	float	time;
 
 	if ( cls.state != ca_active )
 		return;
