+++ win32/in_win.c
@@ -23,7 +23,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "../client/client.h"
 #include "winquake.h"
 
-extern	double	sys_msg_time;
+extern	unsigned int	sys_msg_time;
 
 // joystick defines and variables
 // where should defines be moved?
